backend/pyproject.toml
[project]
name = "career-backend"
version = "0.1.0"
description = "Auth and profile service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "motor",
  "pydantic>=2",
  "pydantic-settings",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "python-multipart",
  "orjson",
  "email-validator",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "httpx",
  "ruff",
]


